amended form to take method requests. Can edit existing events
