+++ 19-AdvancedRouting/frontend/src/pages/EditEvent.js
@@ -4,5 +4,5 @@ import EventForm from "../components/EventForm";
 export default function EditEventPage() {
   const { event } = useRouteLoaderData("event-info");
 
-  return <EventForm event={event} />;
+  return <EventForm event={event} method="patch" />;
 }
